Fix: Refactored Code into Multiple Files for ease of reading.
